requirements.txt
Flask==2.0.2
PyJWT==2.3.0
pyparsing==3.0.6
PyQt5==5.15.6
requests==2.26.0
storm==0.25
